:clap: Remove redundant frontend JS
